rename enrichedSurveyData -> enriched-survey-data
